+++ CHANGELOG.md
@@ -9,7 +9,8 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
-* Added session metadata property. #602
+* Added session metadata property (`Session.metadata`). #602
+* Added session start timestamp (`SessionState.start`) and last activity timestamp (`SessionState.lastActivity`) to the public API. #602
 
 ## [2.2.0] - 2026-03-11